Verbesserungen am Code (einzelne Files gelöscht und gesamten Code richtig in calculator.js eingefügt) und  Übung2 - Erweitere die Tests
