+++ RubinaWeinzettl/02_Probleme_loesen_von_Anforderungen_zu_Loesungen/Tag 3 -  Einfuehrung in TDD/calculator.js
@@ -1,6 +1,22 @@
 function add(a, b) {
     return a + b;
 }
+
+function subtract(a, b) {
+    return a - b;
+}
+
+function multiply(a, b) {
+    return a * b;
+}
+
+function divide(a, b) {
+    return a / b;
+}
+
+// additional function
+function rest(a, b) {
+    return a % b;
+}
   
-module.exports = add;
-  
+module.exports = { add, subtract, divide, multiply, rest };
