--- RubinaWeinzettl/02_Probleme_loesen_von_Anforderungen_zu_Loesungen/Tag 3 -  Einfuehrung in TDD/divide.js
@@ -1,5 +1,0 @@
-function divide(a, b) {
-    return a / b;
-}
-  
-module.exports = divide;
