--- RubinaWeinzettl/02_Probleme_loesen_von_Anforderungen_zu_Loesungen/Tag 3 -  Einfuehrung in TDD/multiply.js
@@ -1,5 +1,0 @@
-function multiply(a, b) {
-    return a * b;
-}
-  
-module.exports = multiply;
